# Examen/doc swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"paths":{"/vols/{id}":{"get":{"tags":["vol-controller"],"operationId":"getVol","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VolDTO"}}}}}}},"/vols/search":{"get":{"tags":["vol-controller"],"operationId":"searchCustomer","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VolDTO"}}}}}}}},"/vols/all":{"get":{"tags":["vol-controller"],"operationId":"Vols","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VolDTO"}}}}}}}},"/reservations/{id}":{"get":{"tags":["reservation-controller"],"operationId":"getReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}},"/reservations/all":{"get":{"tags":["reservation-controller"],"operationId":"Reservations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}}},"/clients/{id}":{"get":{"tags":["vol-controller"],"operationId":"getClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}}}}}},"/clients/all":{"get":{"tags":["vol-controller"],"operationId":"Clients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientDTO"}}}}}}}}},"components":{"schemas":{"AeroportDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"positionGeographique":{"type":"string","enum":["longitude","latitude","altitude"]},"volDTO":{"$ref":"#/components/schemas/VolDTO"},"villeDTO":{"$ref":"#/components/schemas/VilleDTO"}}},"AvionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"nmrPlace":{"type":"integer","format":"int32"},"volDTO":{"$ref":"#/components/schemas/VolDTO"}}},"ClientDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"prenom":{"type":"string"},"email":{"type":"string"}}},"PaysDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"villeDTO":{"$ref":"#/components/schemas/VilleDTO"}}},"ReservationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"},"siteReservation":{"type":"string"},"clientDTO":{"$ref":"#/components/schemas/ClientDTO"},"voyageDTO":{"$ref":"#/components/schemas/VoyageDTO"}}},"VilleDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"aeroportDTO":{"$ref":"#/components/schemas/AeroportDTO"},"paysDTO":{"$ref":"#/components/schemas/PaysDTO"}}},"VolDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"dateDepart":{"type":"string","format":"date-time"},"dateArrivee":{"type":"string","format":"date-time"},"voyageDTO":{"$ref":"#/components/schemas/VoyageDTO"},"avionDTO":{"$ref":"#/components/schemas/AvionDTO"},"aeroportDTO":{"$ref":"#/components/schemas/AeroportDTO"}}},"VoyageDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"reservationDTO":{"$ref":"#/components/schemas/ReservationDTO"}}}}}}